requirements.txt
streamlit
requests
pillow
gtts
numpy
streamlit-camera-input